requirements.txt
fastapi==0.111.0
uvicorn==0.30.3
pyodbc==5.1.0
pandas==2.2.2
python-dotenv==1.0.1
pydantic==2.8.2
